Add support for filtering datasets by category in from_datasets kind
Beginning with `evaluate` steps we need to be able to generate tasks that consume a subset of datasets matching a given category. We don't categorize datasets in `ci/config.yml` (that's just a record of all supported datasets), so this information must be pulled from the parameters instead (as we do in other transforms for other runtime parameters).

This change also moves the `substitute` helper function elsewhere, to make it more re-usable.
